# modules/apps/data-engine/data-engine-rest-impl/rest-openapi.yaml
components:
    schemas:
...
        DataRecord:
            description:
                https://www.schema.org/DataRecord
            properties:
                dataRecordCollectionId:
                    format: int64
                    type: integer
                dataRecordValues:
                    additionalProperties:
                        type: object
                    type: object
                id:
                    format: int64
                    type: integer
...
info:
    description:
        "A Java client JAR is available for use with the group ID 'com.liferay', artifact ID
        'com.liferay.data.engine.rest.client', and version '2.0.19'."
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: "Data Engine"
    version: v2.0
openapi: 3.0.1
paths:
...
    "/data-definitions/{dataDefinitionId}/data-records":
        get:
            operationId: getDataDefinitionDataRecordsPage
            parameters:
                - in: path
                  name: dataDefinitionId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: dataListViewId
                  required: false
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: keywords
                  schema:
                      type: string
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
                - in: query
                  name: sort
                  required: false
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/DataRecord"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/DataRecord"
                                type: array
                    description:
                        ""
            tags: ["DataRecord"]
        post:
            operationId: postDataDefinitionDataRecord
            parameters:
                - in: path
                  name: dataDefinitionId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DataRecord"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/DataRecord"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataRecord"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/DataRecord"
                    description:
                        ""
            tags: ["DataRecord"]
...
    "/data-record-collections/{dataRecordCollectionId}/data-records":
        get:
            operationId: getDataRecordCollectionDataRecordsPage
            parameters:
                - in: path
                  name: dataRecordCollectionId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: dataListViewId
                  required: false
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: keywords
                  schema:
                      type: string
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
                - in: query
                  name: sort
                  required: false
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/DataRecord"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/DataRecord"
                                type: array
                    description:
                        ""
            tags: ["DataRecord"]
        post:
            operationId: postDataRecordCollectionDataRecord
            parameters:
                - in: path
                  name: dataRecordCollectionId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DataRecord"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/DataRecord"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataRecord"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/DataRecord"
                    description:
                        ""
            tags: ["DataRecord"]
    "/data-record-collections/{dataRecordCollectionId}/data-records/export":
        get:
            operationId: getDataRecordCollectionDataRecordExport
            parameters:
                - in: path
                  name: dataRecordCollectionId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
            responses:
                204:
                    content:
                        application/json:
                            schema:
                                type: string
                        application/xml:
                            schema:
                                type: string
                    description:
                        ""
            tags: ["DataRecord"]
...
    "/data-records/{dataRecordId}":
        delete:
            operationId: deleteDataRecord
            parameters:
                - in: path
                  name: dataRecordId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                204:
                    content:
                        application/json: {}
                        application/xml: {}
                    description:
                        ""
            tags: ["DataRecord"]
        get:
            operationId: getDataRecord
            parameters:
                - in: path
                  name: dataRecordId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataRecord"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/DataRecord"
                    description:
                        ""
            tags: ["DataRecord"]
        put:
            operationId: putDataRecord
            parameters:
                - in: path
                  name: dataRecordId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DataRecord"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/DataRecord"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataRecord"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/DataRecord"
                    description:
                        ""
            tags: ["DataRecord"]
...